+++ hypervisor/arch/x86/guest/vmcall.c
@@ -9,6 +9,7 @@
 #include <spinlock.h>
 #include <vcpu.h>
 #include <vm.h>
+#include <virq.h>
 #include <acrn_hv_defs.h>
 #include <hypercall.h>
 #include <trace.h>
